+++ Mile.Helpers/Mile.Helpers.cpp
@@ -13,6 +13,9 @@
 #include <dwmapi.h>
 #pragma comment(lib, "dwmapi.lib")
 
+#include <Uxtheme.h>
+#pragma comment(lib, "Uxtheme.lib")
+
 namespace
 {
     static bool IsPrivatePerMonitorSupportExtensionApplicable()
@@ -275,4 +278,18 @@ EXTERN_C HRESULT WINAPI MileEnableImmersiveDarkModeForWindow(
     }
 
     return hr;
+}
+
+EXTERN_C HRESULT WINAPI MileAllowNonClientDefaultDrawingForWindow(
+    _In_ HWND WindowHandle,
+    _In_ BOOL NewPolicy)
+{
+    WTA_OPTIONS Attribute = { 0 };
+    Attribute.dwMask = WTNCA_NODRAWCAPTION | WTNCA_NODRAWICON;
+    Attribute.dwFlags = NewPolicy ? 0 : Attribute.dwMask;
+    return ::SetWindowThemeAttribute(
+        WindowHandle,
+        WTA_NONCLIENT,
+        &Attribute,
+        sizeof(WTA_OPTIONS));
 }